requirements.txt
aiohttp >= 3.8.1
pytz >= 2022.1
yt-dlp >= 2022.4.8
discord.py[voice] >= 2.3.2
GitPython >= 3.1.27
Pybooru >= 4.2.2
PixivPy-Async >= 1.2.14
PyNaCl >= 1.5.0
psutil >= 5.9.0
colorlog >= 6.6.0
g4f[all] >= 0.3.2.0
pillow >= 10.3.0